Simplify index.js, fix sourcemaps
Sourcemapping wasn't passing filenames properly in all cases, making some files appear blank in the browser, or showing incorrect paths in the browser.

I rewrote the file in the form used by https://github.com/wearefractal/gulp-coffee and simplified it. I haven't been able to break it with any of my sourcemaps.

JSON support needs to be added back in. I don't use it and wouldn't know how to test it.

Thanks for the sourcemapping support!

P.S. The latest master of LiveScript breaks gulp-livescript by returning a string from ast.compile-root, instead of an object. That is not addressed here.